Rename version file for release please
