+++ .github/workflows/release.yml
@@ -15,3 +15,4 @@ jobs:
         with:
           release-type: python
           target-branch: ${{ github.ref_name }}
+          # component: netconfparser #not needed since single component
