+++ src/Display.py
@@ -15,7 +15,7 @@ from Logger import logger
 import os
 import lzma
 import webbrowser
-from __version__ import __version__
+from src.version import __version__
 import Utils
 
 APP_NAME = f"NetConfParser - {__version__}{ " - DEV" if Utils.is_dev_mode() else ""}"
